Add support for ^1.2.3 and ~1.2.3 versions
It adds support for wildcard, caret and tilde versions
for both scoped and unscoped packages so this module could
be used to test most versions found in package.json dependencies:

* 1 - get the latest 1.x.x
* 1.2 - get the latest 1.2.x
* ^1.2.3 - get the latest 1.x.x but at least 1.2.3
* ~1.2.3 - get the latest 1.2.x but at least 1.2.3

But it keeps downloading the entire data for all versions
from the npm registry - see comments to PR #16 at
https://github.com/sindresorhus/package-json/pull/16

It adds tests for the new version formats and makes
sure that the data returned after this pull request
is still the same as before.

It adds examples to the readme.